Move address, city, state and postalCode to a region address fieldset file
